# backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: RefactorForge API
  description: |
    RefactorForge is an advanced code intelligence platform that provides comprehensive repository analysis, 
    pattern extraction, and automated refactoring recommendations. The API offers real-time GitHub scanning, 
    performance monitoring, and intelligent code improvement suggestions.

    ## Key Features
    - **Multi-Repository Analysis**: Simultaneous scanning and analysis of multiple GitHub repositories
    - **Pattern Recognition**: Advanced pattern extraction from TypeScript, JavaScript, Python, and other languages
    - **Security Analysis**: Automated detection of security vulnerabilities and misconfigurations
    - **Performance Optimization**: Identification of performance bottlenecks and optimization opportunities
    - **Type Safety Analysis**: Detection of type safety issues and recommendations for improvements
    - **Real-Time Monitoring**: Live GitHub webhook integration for continuous analysis
    - **Recommendation Engine**: AI-powered suggestions for code improvements and refactoring

    ## Authentication
    Most endpoints require authentication via GitHub token for accessing private repositories and higher rate limits.

    ## Rate Limiting
    - **Authenticated**: 5000 requests/hour
    - **Unauthenticated**: 60 requests/hour

    ## Performance Features
    - Connection pooling and request batching
    - Intelligent caching with configurable TTL
...
  version: "2.1.0"
  contact:
...
servers:
...
security:
  - GitHubToken: []
paths:
...
  /api/repositories:
    get:
      summary: List repositories
      description: |
        Retrieves a paginated list of analyzed repositories with their metadata,
        pattern counts, and analysis status.
      operationId: listRepositories
      tags:
        - Repositories
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of repositories per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: language
          in: query
          description: Filter by programming language
          required: false
          schema:
            type: string
            example: TypeScript
        - name: category
          in: query
          description: Filter by repository category
          required: false
          schema:
            type: string
            example: frontend
      responses:
        '200':
          description: Repositories retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  repositories:
                    type: array
                    items:
                      $ref: '#/components/schemas/Repository'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      summary: Add repository for analysis
      description: |
        Adds a new GitHub repository to the analysis system. The repository will be
        scanned for patterns, security issues, type safety problems, and performance
        bottlenecks.
      operationId: addRepository
      tags:
        - Repositories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddRepositoryRequest'
            example:
              owner: octocat
              repo: Hello-World
              branch: main
              categories: [sample, demo]
      responses:
        '201':
          description: Repository added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Repository'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Repository already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/patterns:
    get:
      summary: Search patterns across repositories
      description: |
        Searches for patterns across all analyzed repositories with advanced
        filtering, sorting, and aggregation options.
      operationId: searchPatterns
      tags:
        - Patterns
      parameters:
        - name: search
          in: query
          description: Search query for pattern content
          required: false
          schema:
            type: string
        - name: category
          in: query
          description: Filter by pattern category
          required: false
          schema:
            type: string
        - name: language
          in: query
          description: Filter by programming language
          required: false
          schema:
            type: string
        - name: repository
          in: query
          description: Filter by repository name
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of patterns per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Patterns retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  patterns:
                    type: array
                    items:
                      $ref: '#/components/schemas/Pattern'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  aggregations:
                    type: object
                    properties:
                      byCategory:
                        type: object
                        additionalProperties:
                          type: integer
                      byLanguage:
                        type: object
                        additionalProperties:
                          type: integer
                      byRepository:
                        type: object
                        additionalProperties:
                          type: integer
...
  /api/github/integrations:
    get:
      summary: List GitHub integrations
      description: |
        Retrieves all configured GitHub repository integrations including
        synchronization status, webhook configuration, and pattern counts.
      operationId: listGitHubIntegrations
      tags:
        - GitHub
      responses:
        '200':
          description: GitHub integrations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  integrations:
                    type: array
                    items:
                      $ref: '#/components/schemas/GitHubIntegration'
                  total:
                    type: integer
                  active:
                    type: integer
                  totalPatterns:
                    type: integer
                  webhooksActive:
                    type: integer

    post:
      summary: Add GitHub integration
      description: |
        Adds a new GitHub repository integration with webhook configuration
        and automatic pattern extraction settings.
      operationId: addGitHubIntegration
      tags:
        - GitHub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                repository:
                  type: string
                  description: Repository name (owner/repo)
                  example: octocat/Hello-World
                settings:
                  type: object
                  properties:
                    autoSave:
                      type: boolean
                      default: true
                    categories:
                      type: array
                      items:
                        type: string
                    minStars:
                      type: integer
                      minimum: 0
                      default: 1
      responses:
        '201':
          description: GitHub integration added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitHubIntegration'
  /api/webhooks/github:
    post:
      summary: GitHub webhook endpoint
      description: |
        Handles GitHub webhook events for real-time repository analysis.
        Supports push, pull request, and issue events for automatic
        pattern extraction and recommendation updates.
      operationId: handleGitHubWebhook
      tags:
        - Webhooks
      parameters:
        - name: X-GitHub-Event
          in: header
          description: GitHub event type
          required: true
          schema:
            type: string
            enum: [push, pull_request, issues, repository]
        - name: X-Hub-Signature-256
          in: header
          description: GitHub webhook signature
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: GitHub webhook payload (varies by event type)
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  processed:
                    type: object
        '400':
          description: Invalid webhook payload
        '401':
          description: Invalid webhook signature
...
components:
  securitySchemes:
    GitHubToken:
      type: http
      scheme: bearer
      bearerFormat: GitHub Personal Access Token
      description: |
        GitHub Personal Access Token for accessing private repositories and higher rate limits.
        Generate at: https://github.com/settings/tokens
  schemas:
...
    Repository:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        fullName:
          type: string
        description:
          type: string
        language:
          type: string
        patternsCount:
          type: integer
        lastUpdated:
          type: string
          format: date-time
        lastAnalyzed:
          type: string
          format: date-time
        categories:
          type: array
          items:
            type: string
        techStack:
          type: array
          items:
            type: string
        metrics:
          type: object
          properties:
            linesOfCode:
              type: integer
            fileCount:
              type: integer
            complexity:
              type: number
    AddRepositoryRequest:
      type: object
      required:
        - owner
        - repo
      properties:
        owner:
          type: string
          description: GitHub username or organization
        repo:
          type: string
          description: Repository name
        branch:
          type: string
          description: Git branch to analyze
          default: main
        categories:
          type: array
          items:
            type: string
          description: Repository categories for classification
    Pattern:
      type: object
      properties:
        id:
          type: string
        repositoryId:
          type: string
        patternType:
          type: string
        patternName:
          type: string
        content:
          type: string
          description: Code snippet or pattern content
        description:
          type: string
        category:
          type: string
        subcategory:
          type: string
        filePath:
          type: string
        lineStart:
          type: integer
        lineEnd:
          type: integer
        language:
          type: string
        framework:
          type: string
        confidenceScore:
          type: number
          minimum: 0
          maximum: 1
        tags:
          type: array
          items:
            type: string
        contextBefore:
          type: string
        contextAfter:
          type: string
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
...
    GitHubIntegration:
      type: object
      properties:
        id:
          type: string
        repository:
          type: string
        status:
          type: string
          enum: [active, inactive, syncing, error]
        lastSync:
          type: string
          format: date-time
        patternsCount:
          type: integer
        webhooksEnabled:
          type: boolean
        branches:
          type: array
          items:
            type: string
        settings:
          type: object
          properties:
            autoSave:
              type: boolean
            categories:
              type: array
              items:
                type: string
            minStars:
              type: integer
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: boolean
          example: true
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
...